% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_suspects.R
\name{check_suspects}
\alias{check_suspects}
\title{Prepares data}
\usage{
check_suspects(dir)
}
\arguments{
\item{dir}{A directory containing results}
}
\value{
A vector of file names that are suspect
}
\description{
Function to prepare data for an object to be taken in to Gibbs sampler
}
