% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampling.R
\name{gibbs_sampling}
\alias{gibbs_sampling}
\title{Gibbs Sampling Function}
\usage{
gibbs_sampling(
  gdata,
  n_iter = 1000,
  burn_in = 500,
  Z_cov_dense = FALSE,
  run_selection_gibbs = FALSE
)
}
\arguments{
\item{gdata}{An object produced by augsynth "multisynth"
gdata$y A numeric vector of responses.
gdata$X A design matrix.
gdata$W A weight matrix.
gdata$Z A covariate matrix.}

\item{n_iter}{Number of iterations to run.}

\item{burn_in}{Number of iterations to discard as burn-in.}

\item{Z_cov_dense}{logical - whether use dense covariance}

\item{run_selection_gibbs}{logical - whether to run selection eqn}
}
\value{
A list of MCMC samples.
}
\description{
This function runs a Gibbs sampler for a given set of parameters.
}
