% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initialize_dense_covariance}
\alias{initialize_dense_covariance}
\title{Initialize a Dense Covariance Matrix Using Cholesky Decomposition}
\usage{
initialize_dense_covariance(p, scale = 1)
}
\arguments{
\item{p}{An integer specifying the dimension of the covariance matrix (i.e.,
the number of rows/columns).}

\item{scale}{A numeric value to scale the resulting covariance matrix. This is
used to control the magnitude of the variances. Default is 1.}
}
\value{
A \eqn{\( p \times p \)} symmetric positive-definite covariance matrix.
}
\description{
This function generates a random \eqn{\( p \times p \)} symmetric positive-definite
covariance matrix using the Cholesky decomposition method. The matrix is
generated by taking \( A^T A \), where \( A \) is a random matrix, ensuring that
the resulting matrix is symmetric and positive-definite.
}
\examples{
set.seed(123)
p <- 3  # Dimension of the covariance matrix
cov_matrix <- initialize_dense_covariance(p)
print(cov_matrix)

}
