% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_kth_column_block_diagonal_fast}
\alias{replace_kth_column_block_diagonal_fast}
\title{Replace k-th column in each block of a block diagonal matrix with CR}
\usage{
replace_kth_column_block_diagonal_fast(M, CR, T, K, N, k)
}
\arguments{
\item{M}{A block diagonal matrix with \code{N} blocks, each of size \verb{T x K}. The matrix is expected
to be in dense format.}

\item{T}{An integer specifying the number of rows in each block of the matrix.}

\item{K}{An integer specifying the number of columns in each block of the matrix.}

\item{N}{An integer specifying the number of blocks in the matrix.}

\item{k}{An integer specifying the index of the column (in each block) to be replaced with the corresponding sub-vector from \code{GRX}.
This value must be in the range \eqn{1 \leq k \leq K}.}

\item{GRX}{A numeric column vector of length \code{N * T} to replace the k-th column in each block of the matrix.}
}
\value{
A matrix \code{M} with the k-th column of each block replaced by the respective sub-vectors from \code{GRX}.
}
\description{
This function replaces the k-th column in each block of a block diagonal matrix \code{M}
with corresponding sub-vectors from a column vector \code{CR}. The matrix \code{M} is assumed
to be composed of \code{N} blocks, each of size \verb{T x K}. The vector \code{CR} is of length
\code{N * T}, and is divided into \code{N} sub-vectors of length \code{T}, each of which replaces
the k-th column of one block in \code{M}.
}
\examples{
# Example of using the replace_kth_column_block_diagonal function
T <- 5  # Number of rows in each block
K <- 3  # Number of columns in each block
N <- 4  # Number of blocks

# Create a block diagonal matrix M with N blocks, each of size T x K
M <- Matrix::bdiag(replicate(N, matrix(rnorm(T * K), T, K), simplify = FALSE))
M <- as.matrix(M)  # Convert the block diagonal matrix to a dense format

# Column vector CR to replace the k-th column in each block (length N * T)
CR <- rnorm(N * T)

# Replace the 2nd column (k = 2) of each block with the corresponding sub-vectors from GRX
k <- 2
M_updated <- replace_kth_column_block_diagonal(M, CR, T, K, N, k)

print(M_updated)  # Print the updated matrix

}
