% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rMVNormCovariance}
\alias{rMVNormCovariance}
\title{Efficient draw from multivariate normal}
\usage{
rMVNormCovariance(n, mu, Sigma)
}
\arguments{
\item{n}{Number of draws.}

\item{mu}{A vector of means.}

\item{Sigma}{A covariance matrix}
}
\description{
Useful for very large draws of a multivariate vector
uses a Cholesky decomposition
}
\examples{
# Example usage
rMVNormCovariance(10, mu = c(0,0),Sigma = diag(2))
}
