% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{winsorize}
\alias{winsorize}
\title{Winsorizing function}
\usage{
winsorize(x, probs = c(0.025, 0.975))
}
\arguments{
\item{x}{A variable name.}

\item{probs}{A numeric list of lower and upper tails}
}
\description{
This function runs a Gibbs sampler for a given set of parameters.
}
