% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepares data}
\usage{
prepare_data(
  dta = NULL,
  res = NULL,
  f.X = as.formula(),
  f.Z = as.formula(),
  flags = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dta}{A data object containing outcomes and covariates}

\item{res}{An object for modified augmented synthetic control results
produced by augsynth multisynth}

\item{f.X}{A formula for X - must include "tvg.dummy" for intervention}

\item{f.Z}{A formula for moderators for intervention HTE}

\item{flags}{A set of needed parameters}

\item{verbose}{Whether to print a lot of information as it's running}
}
\description{
Function to prepare data for an object to be taken in to Gibbs sampler
}
